small ui changes
